cli/operator: extract assembled.start(); thin run() (10c)
Move the servers+start region (metrics + SSV API servers, health prober, contract-event sync, metadata sync, p2p Setup/Start, operatorNode.Start) out of run() into assembled.start(ctx) — a pure move, since it already reads the assembled's fields. run() is now the thin orchestrator: validate, build clients + networkConfig, assemble, defer Close, return start(ctx). The in-goroutine Fatals in start() (metrics/API/startHealthProber/SyncOngoing) remain until the commit-13 errgroup.
